Fix the area change detection
